+++ packages/jsonapi/src/RequestHandling/RequestConstraintTrait.php
@@ -4,12 +4,24 @@ declare(strict_types=1);
 
 namespace EDT\JsonApi\RequestHandling;
 
+use EDT\Querying\Utilities\CollectionConstraintFactory;
 use EDT\Wrapping\Contracts\ContentField;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints as Assert;
 
 trait RequestConstraintTrait
 {
+    private ?CollectionConstraintFactory $collectionConstraintFactory = null;
+
+    private function getCollectionConstraintFactory(): CollectionConstraintFactory
+    {
+        if (null === $this->collectionConstraintFactory) {
+            $this->collectionConstraintFactory = new CollectionConstraintFactory();
+        }
+
+        return $this->collectionConstraintFactory;
+    }
+
     /**
      * Creates a constraint to validate an `id` field value.
      *
@@ -59,10 +71,10 @@ trait RequestConstraintTrait
         return [
             new Assert\NotNull(),
             new Assert\Type('array'),
-            new Assert\Collection([
+            $this->getCollectionConstraintFactory()->exactMatch('relationship references', [
                 ContentField::TYPE => $this->getTypeIdentifierConstraints($typeIdentifier),
                 ContentField::ID => $this->getIdConstraints(null),
-            ], null, null, false, false),
+            ]),
         ];
     }
 }